requirements.txt
Django==4.2.11
djangorestframework==3.15.1
psycopg2-binary==2.9.9
celery==5.3.6
redis==5.0.3
django-celery-beat==2.6.0
python-dotenv==1.0.1
gunicorn==21.2.0
